내역확인
Head: 메인 버튼 삭제 -> 로고 클릭시 메인으로

BodyItem - 예약 버튼 이벤트 추가 -> 예약성공시 마이페이지로 이동

BodyItemList - 가져온 아이템 수에 따라 리스트 Height 자동조절

+++ src/gui/Body.java
@@ -90,7 +90,7 @@ public class Body extends JPanel {
 		repaint();
 	}
 
-	public void showMyPage(int member_key) {
+	public void showMyPage(long member_key) {
 		previous_page = current_page;
 		current_page = new BodyMyPage(this, member_key);
 		removeAll();
